Refactored user leaderboard builder from user into leaderboard
