+++ tracker/leaderboard.py
@@ -1,4 +1,5 @@
 import tracker.db as db
+import tracker.user as user
 
 class Position():
 
@@ -17,3 +18,27 @@ class Position():
             return 'script kiddie'
         else:
             return 'noob'
+
+
+# Leaderboard builder for user entities
+def make_leaderboard(query, args=()):
+    out = []
+    data = db.query_db(query, args)
+    if len(data) == 0: # If no flags found yet send None to template (render empty table)
+        return out
+    pos = 1
+    for d in data:
+        out.append(Position(pos, user.User(d['id'], d['name']), d['score']))
+        pos += 1
+    return out
+
+# Get global leaderboard data
+def get_global():
+    return make_leaderboard('''
+        SELECT u.id, u.name, SUM(f.value) AS score
+        FROM flagsfound ff
+        LEFT JOIN flags f ON f.flag = ff.flag_id
+        LEFT JOIN users u ON u.id = ff.user_id
+        GROUP BY u.id
+        ORDER BY score DESC
+    ''')
